+++ packages/liveblocks-react/src/types.ts
@@ -316,6 +316,7 @@ export type MutationContext<
   S extends LsonObject,
   U extends BaseUserMeta,
 > = {
+  isStorageReady: boolean;
   storage: LiveObject<S>;
   self: User<P, U>;
   others: readonly User<P, U>[];